+++ packages/calcite-components/src/components/dropdown/dropdown.e2e.ts
@@ -22,6 +22,7 @@ import {
 } from "../../tests/utils";
 import type { DropdownItem } from "../dropdown-item/dropdown-item";
 import type { Button } from "../button/button";
+import { ComponentTestTokens, themed } from "../../tests/commonTests/themed";
 import { CSS } from "./resources";
 
 describe("calcite-dropdown", () => {
@@ -39,6 +40,14 @@ describe("calcite-dropdown", () => {
   describe("defaults", () => {
     defaults("calcite-dropdown", [
       {
+        propertyName: "offsetDistance",
+        defaultValue: 0,
+      },
+      {
+        propertyName: "offsetSkidding",
+        defaultValue: 0,
+      },
+      {
         propertyName: "scale",
         defaultValue: "m",
       },
@@ -51,6 +60,14 @@ describe("calcite-dropdown", () => {
 
   describe("reflects", () => {
     reflects("calcite-dropdown", [
+      {
+        propertyName: "offsetDistance",
+        value: 10,
+      },
+      {
+        propertyName: "offsetSkidding",
+        value: 10,
+      },
       {
         propertyName: "scale",
         value: "m",
@@ -137,10 +154,10 @@ describe("calcite-dropdown", () => {
     expect(group1).toEqualAttribute("selection-mode", "multiple");
   });
 
-  it.skip("inheritable non-default props `selectionMode` and `scale` set on parent get passed into items", async () => {
+  it("inheritable non-default props `selectionMode` and `scale` set on parent get passed into items", async () => {
     const page = await newE2EPage();
     await page.setContent(html`
-      <calcite-dropdown selection-mode="single-persist" scale="s">
+      <calcite-dropdown selection-mode="single" scale="s">
         <calcite-button slot="trigger">Open dropdown</calcite-button>
         <calcite-dropdown-group id="group-1">
           <calcite-dropdown-item id="item-1">Content</calcite-dropdown-item>
@@ -149,10 +166,10 @@ describe("calcite-dropdown", () => {
         </calcite-dropdown-group>
       </calcite-dropdown>
     `);
-    const dropdownItems = await findAll(page, "calcite-dropdown-items");
+    const dropdownItems = await findAll(page, "calcite-dropdown-item");
 
     for (const item of dropdownItems) {
-      expect(await item.getProperty("selectionMode")).toBe("single-persist");
+      expect(await item.getProperty("selectionMode")).toBe("single");
       expect(await item.getProperty("scale")).toBe("s");
     }
   });
@@ -1466,5 +1483,19 @@ describe("calcite-dropdown", () => {
       await page.waitForChanges();
       expect(await isElementFocused(page, "#item-2")).toBe(true);
     });
+  });
+
+  describe("theme", () => {
+    const tokens: ComponentTestTokens = {
+      "--calcite-dropdown-width": {
+        targetProp: "inlineSize",
+        shadowSelector: `.${CSS.content}`,
+      },
+      "--calcite-dropdown-background-color": {
+        targetProp: "backgroundColor",
+        shadowSelector: `.${CSS.content}`,
+      },
+    };
+    themed(`<calcite-dropdown open></calcite-dropdown>`, tokens);
   });
 });