+++ packages/calcite-ui-icons/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@4.1.0-next.1...@esri/calcite-ui-icons@4.1.0-next.2) (2025-03-20)
+
+### Features
+
+- add browser join and browser plus ([#11779](https://github.com/Esri/calcite-design-system/issues/11779)) ([44e12d7](https://github.com/Esri/calcite-design-system/commit/44e12d7fb696d7888f987065232451e434052f6a))
+
+## [4.1.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@4.1.0-next.0...@esri/calcite-ui-icons@4.1.0-next.1) (2025-03-19)
+
+### Bug Fixes
+
+- Update README.md ([#11768](https://github.com/Esri/calcite-design-system/issues/11768)) ([7c60d00](https://github.com/Esri/calcite-design-system/commit/7c60d00f73fc0053de2cb70a14d6298d46e8b401))
+
+## [4.1.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@4.0.1-next.1...@esri/calcite-ui-icons@4.1.0-next.0) (2025-03-15)
+
+### Features
+
+- add language-2 ([#11739](https://github.com/Esri/calcite-design-system/issues/11739)) ([5040857](https://github.com/Esri/calcite-design-system/commit/5040857d194c2856142690cb0c47e0497e8d7b30))
+
+## [4.0.1-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@4.0.1-next.0...@esri/calcite-ui-icons@4.0.1-next.1) (2025-03-04)
+
+**Note:** Version bump only for package @esri/calcite-ui-icons
+
 ## [4.0.1-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@4.0.0-next.13...@esri/calcite-ui-icons@4.0.1-next.0) (2025-02-10)
 
 **Note:** Version bump only for package @esri/calcite-ui-icons